Fix ComposerEnvironment storageConfig.bucketRef mapping
In ComposerEnvironment direct reconciler, override autogenerated
StorageConfig_ToProto and StorageConfig_FromProto mappers so that
storageConfig.bucketRef maps to just the GCS bucket name without full
resource path ('projects/../buckets/..'). This resolves API error 400
during ComposerEnvironment creation/update (b/528060386).
